pyproject.toml
[project]
name = "securities-data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "yfinance>=0.2.60",
    "python-dotenv>=1.0.1",
    "duckdb>=1.1.3",
    "s3fs>=2024.10.0",
    "fsspec>=2024.10.0",
    "prefect>=3.1.15",
]

[tool.uv]
dev-dependencies = [
    "fastparquet>=2024.11.0",
    "ipykernel>=6.29.5",
    "minio>=7.2.15",
    "psycopg2-binary>=2.9.11",
    "pytest>=8.3.4",
]
package = true

[tool.setuptools]
packages = ["pipeline"]
